Fix random name generator for python 2 (#3041)

+++ src/azure-cli-testsdk/azure/cli/testsdk/utilities.py
@@ -18,7 +18,7 @@ def create_random_name(prefix='clitest', length=24):
         raise 'The randomized part of the name is shorter than 8, which may not be able to offer ' \
               'enough randomness'
 
-    random_bytes = os.urandom(math.ceil(padding_size / 8) * 5)
+    random_bytes = os.urandom(int(math.ceil(float(padding_size) / 8) * 5))
     random_padding = base64.b32encode(random_bytes)[:padding_size]
 
     return prefix + random_padding.decode().lower()
